Moved isEmpty util to validator-utils.ts and added some other util fns
